refactor(tools): Phase 4 — delete dead agent_zero_mcp, rename agentzero_tools→agentzero, embodied_tool→embodied
Deleted:
  - agent_zero_mcp.py (zero importers, dead HTTP delegate)

Renamed:
  - agentzero_tools.py → agentzero.py
  - embodied_tool.py → embodied.py

Updated imports in:
  - runtime/tools_internal.py, runtime/orchestrator.py, runtime/bridge.py
  - arifosmcp/__init__.py, server.py
  - tools/embodied_instances/arif_mind_reason_embodied.py
  - tests/test_embodied_tool.py

ΔS ≤ 0 | DITEMPA BUKAN DIBERI
