+++ arifosmcp/runtime/agent_registry/cards/arif.json
@@ -1,10 +1,9 @@
 {
-  "agent_id": "arif",
+  "agent_id": "Arif",
   "display_name": "Arif Fazil",
   "version": "human-judge-v1.0",
   "provider": "human",
-  "declared_lane": "AGI",
-  "runtime_lane": "HUMAN",
+  "declared_lane": "HUMAN",
   "federation_role": "sovereign_judge",
   "compliance_tier": "SOVEREIGN",
   "constitutional_floors": 13,
@@ -14,8 +13,6 @@
   "registered_at": "2026-05-08T00:00:00+00:00",
   "sovereign": "Arif Fazil",
   "card_hash": "sha256:arif-sovereign-human-judge-888",
-  "aliases": ["Arif", "arif_fazil"],
-  "canonical_card": "Arif.json",
   "phase": 0,
   "doctrine": "DITEMPA BUKAN DIBERI",
   "verdict": "SEAL",
@@ -23,8 +20,7 @@
   "_meta": {
     "registered_by": "arifOS-kernel",
     "registration_method": "sovereign_binding",
-    "arifos_version": "v2026.05.10-KANON",
-    "registry_file": "runtime/agent_registry/cards/arif.json",
-    "alias_for": "Arif.json"
+    "arifos_version": "v2026.05.08-KANON",
+    "registry_file": "runtime/agent_registry/cards/Arif.json"
   }
 }
