+++ arifosmcp/runtime/orchestrator.py
@@ -222,7 +222,7 @@ async def handle_pns_health(session_id: str) -> RuntimeEnvelope:
 
 async def handle_pns_orchestrate(task: str, session_id: str) -> RuntimeEnvelope:
     """PNS·ORCHESTRATE: Tool routing mediation."""
-    from arifosmcp.tools.agentzero_tools import agentzero_engineer
+    from arifosmcp.tools.agentzero import agentzero_engineer
 
     return await agentzero_engineer(task=task, action_type="execute_code", session_id=session_id)
 
@@ -236,7 +236,7 @@ async def handle_pns_floor(input_data: Any, session_id: str) -> RuntimeEnvelope:
 
 async def handle_pns_redteam(candidate: str, session_id: str) -> RuntimeEnvelope:
     """PNS·REDTEAM: Adversarial testing."""
-    from arifosmcp.tools.agentzero_tools import agentzero_validate
+    from arifosmcp.tools.agentzero import agentzero_validate
 
     return await agentzero_validate(
         input_to_validate=candidate, validation_type="plan", session_id=session_id
