+++ arifosmcp/tools/agentzero.py
