+++ tests/test_embodied_tool.py
@@ -49,7 +49,7 @@ from arifosmcp.schemas.embodied_tool import (
     RiskTier,
     build_embodied_envelope,
 )
-from arifosmcp.tools.embodied_tool import (
+from arifosmcp.tools.embodied import (
     ARIFOS_TOOL_CHARTERS,
     EmbodiedTool,
     register_all_arifos_tools,
